pyproject.toml
[project]
name = "bike-sales-agent"
version = "1.0.0"
description = "PydanticAI-powered bike sales consultant"
dependencies = [
    "pydantic-ai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "openai>=1.100.0",
    "qdrant-client>=1.7.0",
    "redis>=5.0.0",
    "sentence-transformers>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"